# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.AudioMessageRequest:
    description: Mensaje con audio
    properties:
      audio_url:
        example: https://example.com/audio.mp3
        type: string
      caption:
        example: Escucha esto
        type: string
      to:
        example: '@username'
        type: string
    required:
    - audio_url
    - to
    type: object
...
  domain.BulkTextRequest:
    description: Envío masivo de texto a múltiples destinatarios
    properties:
      delay_ms:
        example: 3000
        type: integer
      recipients:
        example:
        - '@user1'
        - '@user2'
        - "+573001234567"
        items:
          type: string
        minItems: 1
        type: array
      text:
        example: Mensaje para todos
        type: string
    required:
    - recipients
    - text
    type: object
...
  domain.FileMessageRequest:
    description: Mensaje con archivo/documento
    properties:
      caption:
        example: Documento adjunto
        type: string
      file_url:
        example: https://example.com/doc.pdf
        type: string
      to:
        example: '@username'
        type: string
    required:
    - file_url
    - to
    type: object
...
  domain.MessageJob:
    description: Estado detallado del mensaje
    properties:
      caption:
        type: string
      created_at:
        type: string
      error:
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      media_url:
        type: string
      send_at:
        type: string
      sent_at:
        type: string
      session_id:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/domain.MessageStatus'
        example: sent
      text:
        type: string
      to:
        example: '@username'
        type: string
      type:
        allOf:
        - $ref: '#/definitions/domain.MessageType'
        example: text
    type: object
  domain.MessageResponse:
    description: Respuesta de envío de mensaje
    properties:
      job_id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      message:
        example: Mensaje en cola
        type: string
      send_at:
        type: string
      status:
        allOf:
        - $ref: '#/definitions/domain.MessageStatus'
        example: pending
    type: object
  domain.MessageStatus:
    enum:
    - pending
    - scheduled
    - sending
    - sent
    - failed
    type: string
    x-enum-varnames:
    - MessageStatusPending
    - MessageStatusScheduled
    - MessageStatusSending
    - MessageStatusSent
    - MessageStatusFailed
  domain.MessageType:
    enum:
    - text
    - photo
    - video
    - audio
    - file
    type: string
    x-enum-varnames:
    - MessageTypeText
    - MessageTypePhoto
    - MessageTypeVideo
    - MessageTypeAudio
    - MessageTypeFile
  domain.PhotoMessageRequest:
    description: Mensaje con foto
    properties:
      caption:
        example: Mira esta imagen
        type: string
      photo_url:
        example: https://example.com/image.jpg
        type: string
      to:
        example: '@username'
        type: string
    required:
    - photo_url
    - to
    type: object
...
  domain.TextMessageRequest:
    description: Mensaje de texto simple
    properties:
      text:
        example: Hola desde la API!
        type: string
      to:
        example: '@username o +573001234567'
        type: string
    required:
    - text
    - to
    type: object
...
  domain.VideoMessageRequest:
    description: Mensaje con video
    properties:
      caption:
        example: Video interesante
        type: string
      to:
        example: '@username'
        type: string
      video_url:
        example: https://example.com/video.mp4
        type: string
    required:
    - to
    - video_url
    type: object
...
  handler.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  handler.MetaResponse:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  handler.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/handler.ErrorResponse'
      meta:
        $ref: '#/definitions/handler.MetaResponse'
      success:
        type: boolean
    type: object
host: localhost:7789
info:
  contact: {}
  description: API REST para gestionar múltiples sesiones de Telegram via MTProto
  title: Telegram API
  version: 0.1.0
paths:
...
  /messages/{jobId}/status:
    get:
      description: Consulta estado de un mensaje enviado
      parameters:
      - description: Job ID
        in: path
        name: jobId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.MessageJob'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Estado del mensaje
      tags:
      - Messages
...
  /sessions/{id}/messages/audio:
    post:
      consumes:
      - application/json
      description: Envía archivo de audio
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Audio
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.AudioMessageRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.MessageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Enviar audio
      tags:
      - Messages
  /sessions/{id}/messages/bulk:
    post:
      consumes:
      - application/json
      description: Envía mensaje de texto a múltiples destinatarios con delay
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Mensaje masivo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.BulkTextRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.MessageResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Envío masivo
      tags:
      - Messages
  /sessions/{id}/messages/file:
    post:
      consumes:
      - application/json
      description: Envía archivo/documento
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Archivo
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.FileMessageRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.MessageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Enviar documento
      tags:
      - Messages
  /sessions/{id}/messages/photo:
    post:
      consumes:
      - application/json
      description: Envía imagen con caption opcional
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Foto
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.PhotoMessageRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.MessageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Enviar foto
      tags:
      - Messages
  /sessions/{id}/messages/text:
    post:
      consumes:
      - application/json
      description: Envía mensaje de texto simple
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Mensaje de texto
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.TextMessageRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.MessageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Enviar texto
      tags:
      - Messages
  /sessions/{id}/messages/video:
    post:
      consumes:
      - application/json
      description: Envía video con caption opcional
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Video
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.VideoMessageRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/domain.MessageResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - BearerAuth: []
      summary: Enviar video
      tags:
      - Messages
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"